\name{bixi_station_departures}
\alias{bixi_station_departures}
\docType{data}
\title{
Daily Departure from BIXI Stations in 2019
}
\description{
These data capture the number of daily departure for 587 bike sharing stations
(rows) through 197 days (columns). The data is limited to the 2019 season of
a Montreal based bike sharing company named BIXI.
}
\usage{data("bixi_station_departures")}
\format{A data frame with 587 rows and 197 columns.}
\source{
Wang, X., Cheng, Z., Trépanier, M., & Sun, L. (2021).
Modeling bike-sharing demand using a regression model with spatially
varying coefficients. Journal of Transport Geography, 93, 103059.
}
\references{
BIXI Montréal (2023). “Open Data.” Accessed: 2023-07-11, URL
\url{https://bixi.com/en/} open-data.
}
\keyword{datasets}
