% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{BKTRBetaPlotMaker}
\alias{BKTRBetaPlotMaker}
\title{R6 class for Creating plots of the BKTR estimates}
\description{
R6 class for creating plots of the BKTR estimates. Can be used
to create plots of the beta estimates for the spatial and temporal features.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-new}{\code{BKTRBetaPlotMaker$new()}}
\item \href{#method-NA-create_temporal_beta_plot}{\code{BKTRBetaPlotMaker$create_temporal_beta_plot()}}
\item \href{#method-NA-get_long_temporal_est_df}{\code{BKTRBetaPlotMaker$get_long_temporal_est_df()}}
\item \href{#method-NA-create_spatial_beta_plots}{\code{BKTRBetaPlotMaker$create_spatial_beta_plots()}}
\item \href{#method-unknown-clone}{\code{BKTRBetaPlotMaker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-new"></a>}}
\if{latex}{\out{\hypertarget{method-NA-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BKTRBetaPlotMaker$new(bktr_regressor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-create_temporal_beta_plot"></a>}}
\if{latex}{\out{\hypertarget{method-NA-create_temporal_beta_plot}{}}}
\subsection{Method \code{create_temporal_beta_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BKTRBetaPlotMaker$create_temporal_beta_plot(
  plot_feature_labels,
  plot_point_label,
  date_format = "\%Y-\%m-\%d"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_long_temporal_est_df"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_long_temporal_est_df}{}}}
\subsection{Method \code{get_long_temporal_est_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BKTRBetaPlotMaker$get_long_temporal_est_df(
  estimates_tensor,
  plot_feature_labels,
  plot_point_label
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-create_spatial_beta_plots"></a>}}
\if{latex}{\out{\hypertarget{method-NA-create_spatial_beta_plots}{}}}
\subsection{Method \code{create_spatial_beta_plots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BKTRBetaPlotMaker$create_spatial_beta_plots(
  plot_feature_labels,
  plot_point_label,
  geo_coordinates,
  nb_cols = 1,
  mapbox_zoom = 9,
  use_dark_mode = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unknown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-unknown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BKTRBetaPlotMaker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
