% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\docType{data}
\name{CompositionOps}
\alias{CompositionOps}
\title{Kernel Composition Operations}
\format{
An object of class \code{list} of length 2.
}
\usage{
CompositionOps
}
\description{
Kernel Composition Operations Enum. Possibilities of operation between
two kernels to generate a new composed kernel.
}
\keyword{datasets}
