% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\docType{data}
\name{DIST_TYPE}
\alias{DIST_TYPE}
\title{Distance Types}
\format{
An object of class \code{list} of length 5.
}
\usage{
DIST_TYPE
}
\description{
Distance Type Enum. Possibilities of distance functions between
two spatial or temporal vectors.
}
\keyword{datasets}
