% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{DistanceCalculator}
\alias{DistanceCalculator}
\title{R6 class for Composed Kernels}
\description{
R6 class for Composed Kernels
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-get_matrix}{\code{DistanceCalculator$get_matrix()}}
\item \href{#method-NA-check_tensor_dimensions}{\code{DistanceCalculator$check_tensor_dimensions()}}
\item \href{#method-NA-calc_euclidean_dist}{\code{DistanceCalculator$calc_euclidean_dist()}}
\item \href{#method-NA-calc_haversine_dist}{\code{DistanceCalculator$calc_haversine_dist()}}
\item \href{#method-NA-calc_manhattan_dist}{\code{DistanceCalculator$calc_manhattan_dist()}}
\item \href{#method-NA-calc_chebyshev_dist}{\code{DistanceCalculator$calc_chebyshev_dist()}}
\item \href{#method-unknown-clone}{\code{DistanceCalculator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_matrix}{}}}
\subsection{Method \code{get_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistanceCalculator$get_matrix(x, distance_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-check_tensor_dimensions"></a>}}
\if{latex}{\out{\hypertarget{method-NA-check_tensor_dimensions}{}}}
\subsection{Method \code{check_tensor_dimensions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistanceCalculator$check_tensor_dimensions(
  x1,
  x2,
  expected_nb_dim,
  expected_last_dim_shape = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-calc_euclidean_dist"></a>}}
\if{latex}{\out{\hypertarget{method-NA-calc_euclidean_dist}{}}}
\subsection{Method \code{calc_euclidean_dist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistanceCalculator$calc_euclidean_dist(x1, x2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-calc_haversine_dist"></a>}}
\if{latex}{\out{\hypertarget{method-NA-calc_haversine_dist}{}}}
\subsection{Method \code{calc_haversine_dist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistanceCalculator$calc_haversine_dist(x1, x2, earth_radius = EARTH_RADIUS_KM)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-calc_manhattan_dist"></a>}}
\if{latex}{\out{\hypertarget{method-NA-calc_manhattan_dist}{}}}
\subsection{Method \code{calc_manhattan_dist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistanceCalculator$calc_manhattan_dist(x1, x2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-calc_chebyshev_dist"></a>}}
\if{latex}{\out{\hypertarget{method-NA-calc_chebyshev_dist}{}}}
\subsection{Method \code{calc_chebyshev_dist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistanceCalculator$calc_chebyshev_dist(x1, x2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unknown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-unknown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DistanceCalculator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
