% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{GeoMercatorProjector}
\alias{GeoMercatorProjector}
\title{Class to project coordinates with mercator projection on a 2D plane}
\description{
Project coordinates with mercator projection on a 2D plane for
a given scale. Keep track of the scale and the center of the projection to
be able to project new coordinates which is useful during interpolation.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeoMercatorProjector-new}{\code{GeoMercatorProjector$new()}}
\item \href{#method-GeoMercatorProjector-project_new_coords}{\code{GeoMercatorProjector$project_new_coords()}}
\item \href{#method-GeoMercatorProjector-clone}{\code{GeoMercatorProjector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoMercatorProjector-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeoMercatorProjector-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoMercatorProjector$new(df, scale = 10)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoMercatorProjector-project_new_coords"></a>}}
\if{latex}{\out{\hypertarget{method-GeoMercatorProjector-project_new_coords}{}}}
\subsection{Method \code{project_new_coords()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoMercatorProjector$project_new_coords(df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeoMercatorProjector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeoMercatorProjector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeoMercatorProjector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
