% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{get_euclidean_dist_tsr}
\alias{get_euclidean_dist_tsr}
\title{Function to compute the euclidean distance between two tensors}
\usage{
get_euclidean_dist_tsr(x1, x2)
}
\description{
Distance Type Enum. Possibilities of distance functions between
two spatial or temporal vectors.
}
