% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_utils.R
\name{get_haversine_dist_matrix}
\alias{get_haversine_dist_matrix}
\title{Function to create a distance matrix from a spatial data matrix}
\usage{
get_haversine_dist_matrix(spatial_matrix)
}
\arguments{
\item{spatial_matrix}{Matrix[numeric]: The matrix of the spatial data containing at least columns
named 'latitude' and 'longitude'}
}
\value{
A new matrix containing the haversine distance between each spatial entity
}
\description{
Function that calculate the distance between each spatial entity used in the BKTR
algorithm using the haversine distance method of calculation.
}
\keyword{internal}
