% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{HyperParamSampler}
\alias{HyperParamSampler}
\title{R6 class for kernel's hyperparameter sampling}
\description{
The HyperParamSampler encapsulate all the behavior related to
the sampling of the kernel hyperparameters
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HyperParamSampler$new()}}
\item \href{#method-set_theta_value}{\code{HyperParamSampler$set_theta_value()}}
\item \href{#method-initialize_theta_bounds}{\code{HyperParamSampler$initialize_theta_bounds()}}
\item \href{#method-prior_fn}{\code{HyperParamSampler$prior_fn()}}
\item \href{#method-sample_rand_theta_value}{\code{HyperParamSampler$sample_rand_theta_value()}}
\item \href{#method-sample}{\code{HyperParamSampler$sample()}}
\item \href{#method-clone}{\code{HyperParamSampler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperParamSampler$new(
  config,
  kernel_generator,
  marginal_ll_eval_fn,
  kernel_hparam_name
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_theta_value"></a>}}
\if{latex}{\out{\hypertarget{method-set_theta_value}{}}}
\subsection{Method \code{set_theta_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperParamSampler$set_theta_value(theta)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initialize_theta_bounds"></a>}}
\if{latex}{\out{\hypertarget{method-initialize_theta_bounds}{}}}
\subsection{Method \code{initialize_theta_bounds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperParamSampler$initialize_theta_bounds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prior_fn"></a>}}
\if{latex}{\out{\hypertarget{method-prior_fn}{}}}
\subsection{Method \code{prior_fn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperParamSampler$prior_fn(theta)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_rand_theta_value"></a>}}
\if{latex}{\out{\hypertarget{method-sample_rand_theta_value}{}}}
\subsection{Method \code{sample_rand_theta_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperParamSampler$sample_rand_theta_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperParamSampler$sample()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HyperParamSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
