% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{KernelComposed}
\alias{KernelComposed}
\title{R6 class for Composed Kernels}
\description{
R6 class for Composed Kernels
}
\keyword{internal}
\section{Super class}{
\code{BKTR::Kernel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-new}{\code{KernelComposed$new()}}
\item \href{#method-NA-core_kernel_fn}{\code{KernelComposed$core_kernel_fn()}}
\item \href{#method-NA-set_distance_matrix}{\code{KernelComposed$set_distance_matrix()}}
\item \href{#method-unknown-clone}{\code{KernelComposed$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="add_jitter_to_kernel"><a href='../../BKTR/html/Kernel.html#method-Kernel-add_jitter_to_kernel'><code>BKTR::Kernel$add_jitter_to_kernel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="get_name"><a href='../../BKTR/html/Kernel.html#method-Kernel-get_name'><code>BKTR::Kernel$get_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="kernel_gen"><a href='../../BKTR/html/Kernel.html#method-Kernel-kernel_gen'><code>BKTR::Kernel$kernel_gen()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-new"></a>}}
\if{latex}{\out{\hypertarget{method-NA-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelComposed$new(left_kernel, right_kernel, new_name, composition_operation)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-core_kernel_fn"></a>}}
\if{latex}{\out{\hypertarget{method-NA-core_kernel_fn}{}}}
\subsection{Method \code{core_kernel_fn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelComposed$core_kernel_fn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-set_distance_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-NA-set_distance_matrix}{}}}
\subsection{Method \code{set_distance_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelComposed$set_distance_matrix(x = NULL, distance_matrix = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unknown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-unknown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelComposed$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
