% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{KernelMatern}
\alias{KernelMatern}
\title{R6 class for Matern Kernels}
\description{
R6 class for Matern Kernels
}
\section{Super class}{
\code{BKTR::Kernel} -> \code{KernelMatern}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KernelMatern-new}{\code{KernelMatern$new()}}
\item \href{#method-KernelMatern-get_smoothness_kernel_fn}{\code{KernelMatern$get_smoothness_kernel_fn()}}
\item \href{#method-KernelMatern-core_kernel_fn}{\code{KernelMatern$core_kernel_fn()}}
\item \href{#method-KernelMatern-clone}{\code{KernelMatern$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="add_jitter_to_kernel"><a href='../../BKTR/html/Kernel.html#method-Kernel-add_jitter_to_kernel'><code>BKTR::Kernel$add_jitter_to_kernel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="kernel_gen"><a href='../../BKTR/html/Kernel.html#method-Kernel-kernel_gen'><code>BKTR::Kernel$kernel_gen()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="plot"><a href='../../BKTR/html/Kernel.html#method-Kernel-plot'><code>BKTR::Kernel$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="set_positions"><a href='../../BKTR/html/Kernel.html#method-Kernel-set_positions'><code>BKTR::Kernel$set_positions()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelMatern-new"></a>}}
\if{latex}{\out{\hypertarget{method-KernelMatern-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelMatern$new(
  smoothness_factor = 1,
  lengthscale = KernelParameter$new(log(2)),
  kernel_variance = 1,
  distance_type = DIST_TYPE$HAVERSINE,
  jitter_value = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelMatern-get_smoothness_kernel_fn"></a>}}
\if{latex}{\out{\hypertarget{method-KernelMatern-get_smoothness_kernel_fn}{}}}
\subsection{Method \code{get_smoothness_kernel_fn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelMatern$get_smoothness_kernel_fn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelMatern-core_kernel_fn"></a>}}
\if{latex}{\out{\hypertarget{method-KernelMatern-core_kernel_fn}{}}}
\subsection{Method \code{core_kernel_fn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelMatern$core_kernel_fn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelMatern-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KernelMatern-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelMatern$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
