% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{KernelParamSampler}
\alias{KernelParamSampler}
\title{R6 class for kernel's hyperparameter sampling}
\description{
The KernelParamSampler encapsulate all the behavior related to
the sampling of the kernel hyperparameters
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KernelParamSampler-new}{\code{KernelParamSampler$new()}}
\item \href{#method-KernelParamSampler-set_theta_value}{\code{KernelParamSampler$set_theta_value()}}
\item \href{#method-KernelParamSampler-initialize_theta_bounds}{\code{KernelParamSampler$initialize_theta_bounds()}}
\item \href{#method-KernelParamSampler-prior_fn}{\code{KernelParamSampler$prior_fn()}}
\item \href{#method-KernelParamSampler-sample_rand_theta_value}{\code{KernelParamSampler$sample_rand_theta_value()}}
\item \href{#method-KernelParamSampler-sample_param}{\code{KernelParamSampler$sample_param()}}
\item \href{#method-KernelParamSampler-sample}{\code{KernelParamSampler$sample()}}
\item \href{#method-KernelParamSampler-clone}{\code{KernelParamSampler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelParamSampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-KernelParamSampler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelParamSampler$new(config, kernel, marginal_ll_eval_fn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelParamSampler-set_theta_value"></a>}}
\if{latex}{\out{\hypertarget{method-KernelParamSampler-set_theta_value}{}}}
\subsection{Method \code{set_theta_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelParamSampler$set_theta_value(theta)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelParamSampler-initialize_theta_bounds"></a>}}
\if{latex}{\out{\hypertarget{method-KernelParamSampler-initialize_theta_bounds}{}}}
\subsection{Method \code{initialize_theta_bounds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelParamSampler$initialize_theta_bounds(param)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelParamSampler-prior_fn"></a>}}
\if{latex}{\out{\hypertarget{method-KernelParamSampler-prior_fn}{}}}
\subsection{Method \code{prior_fn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelParamSampler$prior_fn(param)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelParamSampler-sample_rand_theta_value"></a>}}
\if{latex}{\out{\hypertarget{method-KernelParamSampler-sample_rand_theta_value}{}}}
\subsection{Method \code{sample_rand_theta_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelParamSampler$sample_rand_theta_value(theta_min, theta_max)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelParamSampler-sample_param"></a>}}
\if{latex}{\out{\hypertarget{method-KernelParamSampler-sample_param}{}}}
\subsection{Method \code{sample_param()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelParamSampler$sample_param(param)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelParamSampler-sample"></a>}}
\if{latex}{\out{\hypertarget{method-KernelParamSampler-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelParamSampler$sample()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelParamSampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KernelParamSampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelParamSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
