% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_config.R
\name{KernelSamplerConfig}
\alias{KernelSamplerConfig}
\title{R6 class representing the configuration used for the sampling of a hyperparameter}
\description{
A KernelSamplerConfig object contains all the information necessary to the process
of sampling kernel hyperparameters. (Used in \strong{Algorithm 2} of the paper)
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{slice_sampling_scale}}{The amplitude of the range used for sampling (Paper: \eqn{\rho})}

\item{\code{min_hyper_value}}{The minimal value admissible for the hyperparameter (Paper: \eqn{\phi_{min}})}

\item{\code{max_hyper_value}}{The maximal value admissible for the hyperparameter (Paper: \eqn{\phi_{max}})}

\item{\code{hyper_mu_prior}}{The prior value for the mean of the sampled hyperparameter (Paper: \eqn{\phi})}

\item{\code{hyper_precision_prior}}{The prior value attributed to the precision of the sampled hyperparameter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{KernelSamplerConfig$new()}}
\item \href{#method-clone}{\code{KernelSamplerConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{KernelSamplerConfig} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelSamplerConfig$new(
  slice_sampling_scale = log(10),
  min_hyper_value = log(0.001),
  max_hyper_value = log(1000),
  hyper_mu_prior = 0,
  hyper_precision_prior = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{slice_sampling_scale}}{Numeric: Slice sampling scale for the hyperparameter}

\item{\code{min_hyper_value}}{Numeric: Minimum value of the hyperparameter}

\item{\code{max_hyper_value}}{Numeric: Maximum value of the hyperparameter}

\item{\code{hyper_mu_prior}}{Numeric: Prior mean of the hyperparameter}

\item{\code{hyper_precision_prior}}{Numeric: Prior precision of the hyperparameter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{KernelSamplerConfig} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelSamplerConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
