% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{KernelSE}
\alias{KernelSE}
\title{R6 class for Square Exponential Kernels}
\description{
R6 class for Square Exponential Kernels
}
\section{Super class}{
\code{BKTR::Kernel} -> \code{KernelSE}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KernelSE-new}{\code{KernelSE$new()}}
\item \href{#method-KernelSE-core_kernel_fn}{\code{KernelSE$core_kernel_fn()}}
\item \href{#method-KernelSE-clone}{\code{KernelSE$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="add_jitter_to_kernel"><a href='../../BKTR/html/Kernel.html#method-Kernel-add_jitter_to_kernel'><code>BKTR::Kernel$add_jitter_to_kernel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="kernel_gen"><a href='../../BKTR/html/Kernel.html#method-Kernel-kernel_gen'><code>BKTR::Kernel$kernel_gen()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="plot"><a href='../../BKTR/html/Kernel.html#method-Kernel-plot'><code>BKTR::Kernel$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="set_positions"><a href='../../BKTR/html/Kernel.html#method-Kernel-set_positions'><code>BKTR::Kernel$set_positions()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelSE-new"></a>}}
\if{latex}{\out{\hypertarget{method-KernelSE-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelSE$new(
  lengthscale = KernelParameter$new(log(2)),
  kernel_variance = 1,
  jitter_value = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelSE-core_kernel_fn"></a>}}
\if{latex}{\out{\hypertarget{method-KernelSE-core_kernel_fn}{}}}
\subsection{Method \code{core_kernel_fn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelSE$core_kernel_fn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelSE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KernelSE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelSE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
