% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_utils.R
\name{load_matrix_from_csv}
\alias{load_matrix_from_csv}
\title{A helper function to load CSV data as numeric matrix}
\usage{
load_matrix_from_csv(
  file_location,
  has_header = TRUE,
  columns_to_drop = c(),
  columns_to_keep = NULL,
  rows_to_keep = NULL,
  fill_na_with_zeros = FALSE,
  min_max_normalization = FALSE,
  transpose_matrix = FALSE
)
}
\arguments{
\item{file_location}{String: The file path of the source CSV}

\item{has_header}{Boolean: The new mimimal value for the normalized matrix}

\item{columns_to_drop}{Vector[String]: A vector of column name that can be removed}

\item{columns_to_keep}{Vector[String]: A vector of column name that can remain in the matrix}

\item{rows_to_keep}{Vector[Integer]: A vector of row index that can remain in the matrix}

\item{fill_na_with_zeros}{Boolean: Replace missing values with zeros}

\item{min_max_normalization}{Boolean: Apply a (0-1) min-max normalization on the matrix}

\item{transpose_matrix}{Boolean: Transpose the matrix after all its transformations}
}
\value{
A new matrix on which transformations have been applied
}
\description{
Helper utility function to load and transform CSV data into numeric matrix
that can be loaded as tensor into the BKTR algorithm
}
