% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_likelihoods.R
\name{MarginalLikelihoodEvaluator}
\alias{MarginalLikelihoodEvaluator}
\title{R6 class to evaluate the marginal likelihood of the hyperparameter}
\description{
MarginalLikelihoodEvaluator enable the calculation of the marginal
likelihood of the kernel hyperparameters. This likelihood is used during the sampling
process.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MarginalLikelihoodEvaluator$new()}}
\item \href{#method-calc_likelihood}{\code{MarginalLikelihoodEvaluator$calc_likelihood()}}
\item \href{#method-clone}{\code{MarginalLikelihoodEvaluator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalLikelihoodEvaluator$new(
  rank_decomp,
  nb_covariates,
  covariates,
  omega,
  y,
  is_transposed
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calc_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-calc_likelihood}{}}}
\subsection{Method \code{calc_likelihood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalLikelihoodEvaluator$calc_likelihood(
  kernel_values,
  decomp_values,
  covs_decomp,
  tau
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalLikelihoodEvaluator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
