% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_utils.R
\name{min_max_normalize}
\alias{min_max_normalize}
\title{Function to normalize an R matrix with a min-max method}
\usage{
min_max_normalize(mat, new_min, new_max)
}
\arguments{
\item{mat}{Matrix[numeric]: The matrix on which to apply the normalization}

\item{new_min}{Numeric: The new mimimal value for the normalized matrix}

\item{new_max}{Numeric: The new maximal value for the normalized matrix}
}
\value{
A new matrix normalized via the min-max method
}
\description{
Apply a min-max normalization on a R matrix for a given
new min and new max
}
\keyword{internal}
