% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hyperparams_traceplot}
\alias{plot_hyperparams_traceplot}
\title{Plot Hyperparameters Traceplot}
\usage{
plot_hyperparams_traceplot(
  bktr_reg,
  hyperparameters = NULL,
  show_figure = TRUE,
  fig_width = 9,
  fig_height = 5.5
)
}
\arguments{
\item{bktr_reg}{BKTRRegressor: BKTRRegressor object.}

\item{hyperparameters}{Array or NULL: Array of hyperparameters to plot.
If NULL, plot all hyperparameters. Defaults to NULL.}

\item{show_figure}{Boolean: Whether to show the figure. Defaults to True.}

\item{fig_width}{Integer: Figure width. Defaults to 9.}

\item{fig_height}{Integer: Figure height. Defaults to 5.5.}
}
\value{
ggplot or NULL: ggplot object or NULL if show_figure is set to FALSE.
}
\description{
Plot the evolution of hyperparameters through iterations. (Traceplot)
}
