% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_y_estimates}
\alias{plot_y_estimates}
\title{Plot Y Estimates}
\usage{
plot_y_estimates(
  bktr_reg,
  show_figure = TRUE,
  fig_width = 5,
  fig_height = 5,
  fig_title = "y estimates vs observed y values"
)
}
\arguments{
\item{bktr_reg}{BKTRRegressor: BKTRRegressor object.}

\item{show_figure}{Boolean: Whether to show the figure. Defaults to True.}

\item{fig_width}{Numeric: Figure width when figure is shown. Defaults to 5.}

\item{fig_height}{Numeric: Figure height when figure is shown. Defaults to 5.}

\item{fig_title}{String or NULL: Figure title if provided. Defaults to 'y estimates vs observed y values'}
}
\value{
ggplot or NULL: ggplot object or NULL if show_figure is set to FALSE.
}
\description{
Plot y estimates vs observed y values.
}
