% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{+.Kernel}
\alias{+.Kernel}
\title{Operator overloading for kernel addition}
\usage{
\method{+}{Kernel}(k1, k2)
}
\arguments{
\item{k1}{Kernel: The left kernel to use for composition}

\item{k2}{Kernel: The right kernel to use for composition}
}
\value{
A new \code{KernelAddComposed} object.
}
\description{
Operator overloading for kernel addition
}
