% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bktr.R
\name{print.BKTRRegressor}
\alias{print.BKTRRegressor}
\title{Print the summary of a BKTRRegressor instance}
\usage{
\method{print}{BKTRRegressor}(x, ...)
}
\arguments{
\item{x}{A BKTRRegressor instance}

\item{...}{Additional arguments to comply with generic function}
}
\description{
Print the summary of a BKTRRegressor instance
}
