% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_generator.R
\name{SpatialKernelGenerator}
\alias{SpatialKernelGenerator}
\title{R6 class that generates Spatial Kernels}
\description{
A SpatialKernelGenerator can create spatial kernels according
to provided parameters.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SpatialKernelGenerator$new()}}
\item \href{#method-kernel_gen}{\code{SpatialKernelGenerator$kernel_gen()}}
\item \href{#method-clone}{\code{SpatialKernelGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SpatialKernelGenerator} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialKernelGenerator$new(
  distance_matrix,
  smoothness_factor,
  kernel_variance,
  ini_spatial_length_scale = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distance_matrix}}{Matrix[numeric]: The distance matrix between each spatial entity}

\item{\code{smoothness_factor}}{Integer (1,3,5): Factor deciding the matern kernel's type (3 for 3/2)}

\item{\code{kernel_variance}}{Numeric: The variance used in the kernel}

\item{\code{ini_spatial_length_scale}}{Numeric: Initial value for the spatial length scale}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{SpatialKernelGenerator} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kernel_gen"></a>}}
\if{latex}{\out{\hypertarget{method-kernel_gen}{}}}
\subsection{Method \code{kernel_gen()}}{
Generate and set spatial kernel values using current parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialKernelGenerator$kernel_gen()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new numeric matrix containing the spatial kernel
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialKernelGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
