% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bktr.R
\name{summary.BKTRRegressor}
\alias{summary.BKTRRegressor}
\title{Summarize a BKTRRegressor instance}
\usage{
\method{summary}{BKTRRegressor}(object, ...)
}
\arguments{
\item{object}{A BKTRRegressor instance}

\item{...}{Additional arguments to comply with generic function}
}
\description{
Summarize a BKTRRegressor instance
}
