% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{TauSampler}
\alias{TauSampler}
\title{R6 class for the Tau precision hyperparameter sampling}
\description{
Encapsulate all the behavior that allows to generate new tau values
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TauSampler-new}{\code{TauSampler$new()}}
\item \href{#method-TauSampler-sample}{\code{TauSampler$sample()}}
\item \href{#method-TauSampler-clone}{\code{TauSampler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TauSampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-TauSampler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TauSampler$new(a_0, b_0, nb_observations)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TauSampler-sample"></a>}}
\if{latex}{\out{\hypertarget{method-TauSampler-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TauSampler$sample(total_sq_error)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TauSampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TauSampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TauSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
