% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_generator.R
\name{TemporalKernelGenerator}
\alias{TemporalKernelGenerator}
\title{R6 class that generates Temporal Kernels}
\description{
A TemporalKernelGenerator can create temporal kernels according
to provided parameters. A temporal kernel can use a periodic kernel function, a
squared exponential (SE) kernel function or a mixture of both.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TemporalKernelGenerator$new()}}
\item \href{#method-kernel_gen}{\code{TemporalKernelGenerator$kernel_gen()}}
\item \href{#method-clone}{\code{TemporalKernelGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TemporalKernelGenerator} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporalKernelGenerator$new(
  kernel_fn_name,
  nb_time_segments,
  period_length,
  kernel_variance,
  ini_period_length_scale = NULL,
  ini_decay_time_scale = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{kernel_fn_name}}{String ('periodic', 'se', 'periodic_se'): The kernel function's name}

\item{\code{nb_time_segments}}{Integer: The number of time segments for the kernel (e,g. number of days)}

\item{\code{period_length}}{Integer: The period's length for the kernel (e,g. 7 weekly patterns)}

\item{\code{kernel_variance}}{Numeric: The variance used in the kernel}

\item{\code{ini_period_length_scale}}{Numeric: Initial value for the periodic length scale}

\item{\code{ini_decay_time_scale}}{Numeric: Initial value for the decay time scale}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{TemporalKernelGenerator} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kernel_gen"></a>}}
\if{latex}{\out{\hypertarget{method-kernel_gen}{}}}
\subsection{Method \code{kernel_gen()}}{
Generate and set temporal kernel values using current parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporalKernelGenerator$kernel_gen()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new numeric matrix containing the temporal kernel
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporalKernelGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
