% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_ops.R
\docType{data}
\name{TSR}
\alias{TSR}
\title{Tensor Operator Singleton}
\format{
An object of class \code{TensorOperator} (inherits from \code{Singleton}, \code{R6}) of length 19.
}
\usage{
TSR
}
\description{
Singleton instance of the \code{TensorOperator} class that contains
all informations related the tensor API; tensor methods, used data type and used device.
}
\keyword{datasets}
