% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{BixiData}
\alias{BixiData}
\title{BIXI Data Class}
\description{
R6 class encapsulating the BIXI data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BixiData-new}{\code{BixiData$new()}}
\item \href{#method-BixiData-clone}{\code{BixiData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BixiData-new"></a>}}
\if{latex}{\out{\hypertarget{method-BixiData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BixiData$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BixiData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BixiData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BixiData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
