% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bktr.R
\name{BKTRRegressor}
\alias{BKTRRegressor}
\title{R6 class encapsulating the BKTR regression steps}
\description{
A BKTRRegressor holds all the key elements to accomplish the MCMC sampling
algorithm (\strong{Algorithm 1} of the paper).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BKTRRegressor$new()}}
\item \href{#method-mcmc_sampling}{\code{BKTRRegressor$mcmc_sampling()}}
\item \href{#method-clone}{\code{BKTRRegressor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{BKTRRegressor} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BKTRRegressor$new(
  bktr_config,
  temporal_covariate_matrix,
  spatial_covariate_matrix,
  spatial_distance_matrix,
  y,
  omega
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bktr_config}}{BKTRConfig: Configuration used for the BKTR regression}

\item{\code{temporal_covariate_matrix}}{matrix[numeric]: Temporal Covariates}

\item{\code{spatial_covariate_matrix}}{matrix[numeric]: Spatial Covariates}

\item{\code{spatial_distance_matrix}}{matrix[numeric]: Distance between spatial entities}

\item{\code{y}}{matrix[numeric]: Response variable that we are trying to predict}

\item{\code{omega}}{matrix[boolean]: Mask showing if a y observation is missing or not}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{BKTRRegressor} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mcmc_sampling"></a>}}
\if{latex}{\out{\hypertarget{method-mcmc_sampling}{}}}
\subsection{Method \code{mcmc_sampling()}}{
Launch the MCMC sampling process. \cr
For a predefined number of iterations:
\enumerate{
\item{Sample the spatial length scale hyperparameter}
\item{Sample the decay time scale hyperparameter}
\item{Sample the periodic length scale hyperparameter}
\item{Sample the precision matrix from a wishart distribution}
\item{Sample a new spatial covariate decomposition}
\item{Sample a new covariate decomposition}
\item{Sample a new temporal covariate decomposition}
\item{Calculate respective errors for the iterations}
\item{Sample a new tau value}
\item{Collect all the important data for the iteration}
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BKTRRegressor$mcmc_sampling()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing the results of the MCMC sampling.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BKTRRegressor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
