% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bktr.R
\name{BKTRRegressor}
\alias{BKTRRegressor}
\title{R6 class encapsulating the BKTR regression elements}
\description{
A BKTRRegressor holds all the key elements to accomplish the MCMC sampling
algorithm (\strong{Algorithm 1} of the paper).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BKTRRegressor-new}{\code{BKTRRegressor$new()}}
\item \href{#method-BKTRRegressor-mcmc_sampling}{\code{BKTRRegressor$mcmc_sampling()}}
\item \href{#method-BKTRRegressor-predict}{\code{BKTRRegressor$predict()}}
\item \href{#method-BKTRRegressor-get_iterations_betas}{\code{BKTRRegressor$get_iterations_betas()}}
\item \href{#method-BKTRRegressor-get_beta_summary_df}{\code{BKTRRegressor$get_beta_summary_df()}}
\item \href{#method-BKTRRegressor-clone}{\code{BKTRRegressor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BKTRRegressor-new"></a>}}
\if{latex}{\out{\hypertarget{method-BKTRRegressor-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{BKTRRegressor} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BKTRRegressor$new(
  data_df,
  spatial_positions_df,
  temporal_positions_df,
  rank_decomp = 10,
  burn_in_iter = 500,
  sampling_iter = 500,
  formula = NULL,
  spatial_kernel = KernelMatern$new(smoothness_factor = 3),
  temporal_kernel = KernelSE$new(),
  sigma_r = 0.01,
  a_0 = 1e-06,
  b_0 = 1e-06,
  has_geo_coords = TRUE,
  geo_coords_scale = 10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_df}}{data.table: A dataframe containing all the covariates
through time and space. It is important that the dataframe has a two
indexes named `location` and `time` respectively. The dataframe should
also contain every possible combinations of `location` and `time`
(i.e. even missing rows should be filled present but filled with NaN).
So if the dataframe has 10 locations and 5 time points, it should have
50 rows (10 x 5). If formula is None, the dataframe should contain
the response variable `Y` as the first column. Note that the covariate
columns cannot contain NaN values, but the response variable can.}

\item{\code{spatial_positions_df}}{data.table: Spatial kernel input tensor used
to calculate covariates' distance. Vector of length equal to the number of location points.}

\item{\code{temporal_positions_df}}{data.table: Temporal kernel input tensor used to
calculate covariate distance. Vector of length equal to the number of time points.}

\item{\code{rank_decomp}}{Integer: Rank of the CP decomposition (Paper -- \eqn{R}). Defaults to 10.}

\item{\code{burn_in_iter}}{Integer: Number of iteration before sampling (Paper -- :math:`K_1`). Defaults to 500.}

\item{\code{sampling_iter}}{Integer: Number of sampling iterations (Paper -- :math:`K_2`). Defaults to 500.}

\item{\code{formula}}{A Wilkinson R formula to specify the relation
between the response variable `Y` and the covariates. If Null, the first
column of the data frame will be used as the response variable and all the
other columns will be used as the covariates.  Defaults to Null.}

\item{\code{spatial_kernel}}{Kernel: Spatial kernel Used. Defaults to
a KernelMatern(smoothness_factor=3).}

\item{\code{temporal_kernel}}{Kernel: Temporal kernel used. Defaults to KernelSE().}

\item{\code{sigma_r}}{Numeric:  Variance of the white noise process (\eqn{\tau^{-1}})
defaults to 1E-2.}

\item{\code{a_0}}{Numeric: Initial value for the shape (\eqn{\alpha}) in the gamma function
generating tau defaults to 1E-6.}

\item{\code{b_0}}{Numeric: Initial value for the rate (\eqn{\beta}) in the gamma function
generating tau defaults to 1E-6.}

\item{\code{has_geo_coords}}{Boolean: Whether the spatial positions df use geographic coordinates
(latitude, longitude). Defaults to TRUE.}

\item{\code{geo_coords_scale}}{Numeric: Scale factor to convert geographic coordinates to euclidean
2D space via Mercator projection using x & y domains of [-scale/2, +scale/2]. Only used if
has_geo_coords is TRUE. Defaults to 10.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{BKTRRegressor} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BKTRRegressor-mcmc_sampling"></a>}}
\if{latex}{\out{\hypertarget{method-BKTRRegressor-mcmc_sampling}{}}}
\subsection{Method \code{mcmc_sampling()}}{
Launch the MCMC sampling process. \cr
For a predefined number of iterations:
\enumerate{
\item{Sample spatial kernel hyperparameters}
\item{Sample temporal kernel hyperparameters}
\item{Sample the precision matrix from a wishart distribution}
\item{Sample a new spatial covariate decomposition}
\item{Sample a new feature covariate decomposition}
\item{Sample a new temporal covariate decomposition}
\item{Calculate respective errors for the iterations}
\item{Sample a new tau value}
\item{Collect all the important data for the iteration}
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BKTRRegressor$mcmc_sampling()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing the results of the MCMC sampling.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BKTRRegressor-predict"></a>}}
\if{latex}{\out{\hypertarget{method-BKTRRegressor-predict}{}}}
\subsection{Method \code{predict()}}{
Use interpolation to predict betas and response values for new data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BKTRRegressor$predict(
  new_data_df,
  new_spatial_positions_df = NULL,
  new_temporal_positions_df = NULL,
  jitter = 1e-05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data_df}}{data.table: New covariates. Must have the same columns as
the covariates used to fit the model. The index should contain the combination
of all old spatial coordinates with all new temporal coordinates, the combination
of all new spatial coordinates with all old temporal coordinates, and the
combination of all new spatial coordinates with all new temporal coordinates.}

\item{\code{new_spatial_positions_df}}{data.table or NULL: A data frame containing the new
spatial positions. Defaults to NULL.}

\item{\code{new_temporal_positions_df}}{data.table or NULL: A data frame containing the new
temporal positions. Defaults to NULL.}

\item{\code{jitter}}{Numeric or NULL: A small value to add to the diagonal of the precision matrix.
Defaults to NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List: A list of two dataframes. The first represents the beta
  forecasted for all new spatial locations or temporal points.
  The second represents the forecasted response for all new spatial
  locations or temporal points.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BKTRRegressor-get_iterations_betas"></a>}}
\if{latex}{\out{\hypertarget{method-BKTRRegressor-get_iterations_betas}{}}}
\subsection{Method \code{get_iterations_betas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BKTRRegressor$get_iterations_betas(
  spatial_label,
  temporal_label,
  feature_label
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BKTRRegressor-get_beta_summary_df"></a>}}
\if{latex}{\out{\hypertarget{method-BKTRRegressor-get_beta_summary_df}{}}}
\subsection{Method \code{get_beta_summary_df()}}{
Get a summary of estimated beta values. If no labels are given,
then the summary is for all the betas. If labels are given, then the summary
is for the given labels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BKTRRegressor$get_beta_summary_df(
  spatial_labels,
  temporal_labels,
  feature_labels
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{spatial_labels}}{vector: The spatial labels to get the summary for the summary.}

\item{\code{temporal_labels}}{vector: The temporal labels to get the summary for the summary.}

\item{\code{feature_labels}}{vector: The feature labels to get the summary for the summary.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new data.table with the summary for the given label.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BKTRRegressor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BKTRRegressor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BKTRRegressor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
