% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{check_dist_tensor_dimensions}
\alias{check_dist_tensor_dimensions}
\title{Distance Types}
\usage{
check_dist_tensor_dimensions(
  x1,
  x2,
  expected_nb_dim = 2,
  expected_last_dim_shape = NULL
)
}
\description{
Distance Type Enum. Possibilities of distance functions between
two spatial or temporal vectors.
}
