% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{get_euclidean_dist_tsr}
\alias{get_euclidean_dist_tsr}
\title{Function to compute a tensor's euclidean distance}
\usage{
get_euclidean_dist_tsr(x)
}
\description{
Function to compute the euclidean distance between a tensor and its transpose.
}
