% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{KernelComposed}
\alias{KernelComposed}
\title{R6 class for Composed Kernels}
\description{
R6 class for Composed Kernels
}
\keyword{internal}
\section{Super class}{
\code{BKTR::Kernel} -> \code{KernelComposed}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KernelComposed-new}{\code{KernelComposed$new()}}
\item \href{#method-KernelComposed-core_kernel_fn}{\code{KernelComposed$core_kernel_fn()}}
\item \href{#method-KernelComposed-set_positions}{\code{KernelComposed$set_positions()}}
\item \href{#method-KernelComposed-clone}{\code{KernelComposed$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="add_jitter_to_kernel"><a href='../../BKTR/html/Kernel.html#method-Kernel-add_jitter_to_kernel'><code>BKTR::Kernel$add_jitter_to_kernel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="kernel_gen"><a href='../../BKTR/html/Kernel.html#method-Kernel-kernel_gen'><code>BKTR::Kernel$kernel_gen()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="plot"><a href='../../BKTR/html/Kernel.html#method-Kernel-plot'><code>BKTR::Kernel$plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelComposed-new"></a>}}
\if{latex}{\out{\hypertarget{method-KernelComposed-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelComposed$new(left_kernel, right_kernel, new_name, composition_operation)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelComposed-core_kernel_fn"></a>}}
\if{latex}{\out{\hypertarget{method-KernelComposed-core_kernel_fn}{}}}
\subsection{Method \code{core_kernel_fn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelComposed$core_kernel_fn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelComposed-set_positions"></a>}}
\if{latex}{\out{\hypertarget{method-KernelComposed-set_positions}{}}}
\subsection{Method \code{set_positions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelComposed$set_positions(positions_df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelComposed-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KernelComposed-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelComposed$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
