% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{KernelMatern}
\alias{KernelMatern}
\title{R6 class for Matern Kernels}
\description{
R6 class for Matern Kernels
}
\examples{
\dontshow{if (torch::torch_is_installed()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Create a new Matern 3/2 kernel
k_matern <- KernelMatern$new(smoothness_factor = 3)
# Set the kernel's positions
positions_df <- data.frame(x=c(-4, 0, 3), y=c(-2, 0, 2))
k_matern$set_positions(positions_df)
# Generate the kernel's covariance matrix
k_matern$kernel_gen()
\dontshow{\}) # examplesIf}
}
\section{Super class}{
\code{\link[BKTR:Kernel]{BKTR::Kernel}} -> \code{KernelMatern}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lengthscale}}{The lengthscale parameter instance of the kernel}

\item{\code{smoothness_factor}}{The smoothness factor of the kernel}

\item{\code{has_dist_matrix}}{Identify if the kernel has a distance matrix or not}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KernelMatern-new}{\code{KernelMatern$new()}}
\item \href{#method-KernelMatern-get_smoothness_kernel_fn}{\code{KernelMatern$get_smoothness_kernel_fn()}}
\item \href{#method-KernelMatern-core_kernel_fn}{\code{KernelMatern$core_kernel_fn()}}
\item \href{#method-KernelMatern-clone}{\code{KernelMatern$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="add_jitter_to_kernel"><a href='../../BKTR/html/Kernel.html#method-Kernel-add_jitter_to_kernel'><code>BKTR::Kernel$add_jitter_to_kernel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="kernel_gen"><a href='../../BKTR/html/Kernel.html#method-Kernel-kernel_gen'><code>BKTR::Kernel$kernel_gen()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="plot"><a href='../../BKTR/html/Kernel.html#method-Kernel-plot'><code>BKTR::Kernel$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="set_positions"><a href='../../BKTR/html/Kernel.html#method-Kernel-set_positions'><code>BKTR::Kernel$set_positions()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelMatern-new"></a>}}
\if{latex}{\out{\hypertarget{method-KernelMatern-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{KernelMatern} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelMatern$new(
  smoothness_factor = 5,
  lengthscale = KernelParameter$new(2),
  kernel_variance = 1,
  jitter_value = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{smoothness_factor}}{Numeric: The smoothness factor of the kernel (1, 3 or 5)}

\item{\code{lengthscale}}{KernelParameter: The lengthscale parameter instance of the kernel}

\item{\code{kernel_variance}}{Numeric: The variance of the kernel}

\item{\code{jitter_value}}{Numeric: The jitter value to add to the kernel matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelMatern-get_smoothness_kernel_fn"></a>}}
\if{latex}{\out{\hypertarget{method-KernelMatern-get_smoothness_kernel_fn}{}}}
\subsection{Method \code{get_smoothness_kernel_fn()}}{
Method to the get the smoothness kernel function for a given integer smoothness factor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelMatern$get_smoothness_kernel_fn()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The smoothness kernel function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelMatern-core_kernel_fn"></a>}}
\if{latex}{\out{\hypertarget{method-KernelMatern-core_kernel_fn}{}}}
\subsection{Method \code{core_kernel_fn()}}{
Method to compute the core kernel's covariance matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelMatern$core_kernel_fn()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The core kernel's covariance matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelMatern-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KernelMatern-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelMatern$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
