% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{KernelPeriodic}
\alias{KernelPeriodic}
\title{R6 class for Periodic Kernels}
\description{
R6 class for Periodic Kernels
}
\section{Super class}{
\code{\link[BKTR:Kernel]{BKTR::Kernel}} -> \code{KernelPeriodic}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lengthscale}}{The lengthscale parameter instance of the kernel}

\item{\code{period_length}}{The period length parameter instance of the kernel}

\item{\code{has_dist_matrix}}{Identify if the kernel has a distance matrix or not}

\item{\code{name}}{The kernel's name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KernelPeriodic-new}{\code{KernelPeriodic$new()}}
\item \href{#method-KernelPeriodic-core_kernel_fn}{\code{KernelPeriodic$core_kernel_fn()}}
\item \href{#method-KernelPeriodic-clone}{\code{KernelPeriodic$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="add_jitter_to_kernel"><a href='../../BKTR/html/Kernel.html#method-Kernel-add_jitter_to_kernel'><code>BKTR::Kernel$add_jitter_to_kernel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="kernel_gen"><a href='../../BKTR/html/Kernel.html#method-Kernel-kernel_gen'><code>BKTR::Kernel$kernel_gen()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="plot"><a href='../../BKTR/html/Kernel.html#method-Kernel-plot'><code>BKTR::Kernel$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BKTR" data-topic="Kernel" data-id="set_positions"><a href='../../BKTR/html/Kernel.html#method-Kernel-set_positions'><code>BKTR::Kernel$set_positions()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelPeriodic-new"></a>}}
\if{latex}{\out{\hypertarget{method-KernelPeriodic-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{KernelPeriodic} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelPeriodic$new(
  lengthscale = KernelParameter$new(2),
  period_length = KernelParameter$new(2),
  kernel_variance = 1,
  jitter_value = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lengthscale}}{KernelParameter: The lengthscale parameter instance of the kernel}

\item{\code{period_length}}{KernelParameter: The period length parameter instance of the kernel}

\item{\code{kernel_variance}}{Numeric: The variance of the kernel}

\item{\code{jitter_value}}{Numeric: The jitter value to add to the kernel matrix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{KernelPeriodic} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelPeriodic-core_kernel_fn"></a>}}
\if{latex}{\out{\hypertarget{method-KernelPeriodic-core_kernel_fn}{}}}
\subsection{Method \code{core_kernel_fn()}}{
Method to compute the core kernel's covariance matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelPeriodic$core_kernel_fn()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The core kernel's covariance matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KernelPeriodic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KernelPeriodic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KernelPeriodic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
