% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_beta_dists}
\alias{plot_beta_dists}
\title{Plot Beta Coefficients Distribution}
\usage{
plot_beta_dists(
  bktr_reg,
  labels_list,
  show_figure = TRUE,
  fig_width = 9,
  fig_height = 6,
  fig_resolution = 200
)
}
\arguments{
\item{bktr_reg}{BKTRRegressor: BKTRRegressor object.}

\item{labels_list}{List: List of labels tuple (spatial, temporal, feature) for
which to plot the beta distribution through iterations}

\item{show_figure}{Boolean: Whether to show the figure. Defaults to True.}

\item{fig_width}{Integer: Figure width. Defaults to 9.}

\item{fig_height}{Integer: Figure height. Defaults to 6.}

\item{fig_resolution}{Numeric: Figure resolution PPI. Defaults to 200.}
}
\value{
ggplot or NULL: ggplot object or NULL if show_figure is set to FALSE.
}
\description{
Plot the distribution of beta values for a given list of labels.
}
