% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_covariates_beta_dists}
\alias{plot_covariates_beta_dists}
\title{Plot Beta Coefficients Distribution Regrouped by Covariates}
\usage{
plot_covariates_beta_dists(
  bktr_reg,
  feature_labels,
  show_figure = TRUE,
  fig_width = 9,
  fig_height = 6
)
}
\arguments{
\item{bktr_reg}{BKTRRegressor: BKTRRegressor object.}

\item{feature_labels}{Array or NULL: Array of feature labels for
which to plot the beta estimates distribution. If NULL plot for all features.}

\item{show_figure}{Boolean: Whether to show the figure. Defaults to True.}

\item{fig_width}{Integer: Figure width. Defaults to 9.}

\item{fig_height}{Integer: Figure height. Defaults to 6.}
}
\value{
ggplot or NULL: ggplot object or NULL if show_figure is set to FALSE.
}
\description{
Plot the distribution of beta estimates regrouped by covariates.
}
