% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_spatial_betas}
\alias{plot_spatial_betas}
\title{Plot Spatial Beta Coefficients}
\usage{
plot_spatial_betas(
  bktr_reg,
  plot_feature_labels,
  temporal_point_label,
  nb_cols = 1,
  use_dark_mode = TRUE,
  show_figure = TRUE,
  fig_width = 8.5,
  fig_height = 5.5
)
}
\arguments{
\item{bktr_reg}{BKTRRegressor: BKTRRegressor object.}

\item{plot_feature_labels}{Array: Array of feature labels to plot.}

\item{temporal_point_label}{String: Temporal point label to plot.}

\item{nb_cols}{Integer: The number of columns to use in the facet grid.}

\item{use_dark_mode}{Boolean: Whether to use a dark mode for the geographic map or not.}

\item{show_figure}{Boolean: Whether to show the figure. Defaults to True.}

\item{fig_width}{Numeric: Figure width when figure is shown. Defaults to 8.5.}

\item{fig_height}{Numeric: Figure height when figure is shown. Defaults to 5.5.}
}
\value{
ggplot or NULL: ggplot object or NULL if show_figure is set to FALSE.
}
\description{
Create a plot of beta values through space for a given
temporal point and a set of feature labels.
}
