% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_logger.R
\name{ResultLogger}
\alias{ResultLogger}
\title{R6 class for Logging BKTR Results}
\description{
The ResultLogger encapsulate all the behavior related to
logging iteration information during the BKTR algorithm
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResultLogger-new}{\code{ResultLogger$new()}}
\item \href{#method-ResultLogger-collect_iter_samples}{\code{ResultLogger$collect_iter_samples()}}
\item \href{#method-ResultLogger-set_error_metrics}{\code{ResultLogger$set_error_metrics()}}
\item \href{#method-ResultLogger-set_y_and_beta_estimates}{\code{ResultLogger$set_y_and_beta_estimates()}}
\item \href{#method-ResultLogger-log_final_iter_results}{\code{ResultLogger$log_final_iter_results()}}
\item \href{#method-ResultLogger-summary}{\code{ResultLogger$summary()}}
\item \href{#method-ResultLogger-get_beta_summary_df}{\code{ResultLogger$get_beta_summary_df()}}
\item \href{#method-ResultLogger-get_iteration_betas_tensor}{\code{ResultLogger$get_iteration_betas_tensor()}}
\item \href{#method-ResultLogger-clone}{\code{ResultLogger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultLogger-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResultLogger-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultLogger$new(
  y,
  omega,
  covariates,
  nb_burn_in_iter,
  nb_sampling_iter,
  rank_decomp,
  formula,
  spatial_labels,
  temporal_labels,
  feature_labels,
  spatial_kernel,
  temporal_kernel
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultLogger-collect_iter_samples"></a>}}
\if{latex}{\out{\hypertarget{method-ResultLogger-collect_iter_samples}{}}}
\subsection{Method \code{collect_iter_samples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultLogger$collect_iter_samples(iter, tau_value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultLogger-set_error_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-ResultLogger-set_error_metrics}{}}}
\subsection{Method \code{set_error_metrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultLogger$set_error_metrics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultLogger-set_y_and_beta_estimates"></a>}}
\if{latex}{\out{\hypertarget{method-ResultLogger-set_y_and_beta_estimates}{}}}
\subsection{Method \code{set_y_and_beta_estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultLogger$set_y_and_beta_estimates(decomp_tensors_map, iter)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultLogger-log_final_iter_results"></a>}}
\if{latex}{\out{\hypertarget{method-ResultLogger-log_final_iter_results}{}}}
\subsection{Method \code{log_final_iter_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultLogger$log_final_iter_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultLogger-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ResultLogger-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultLogger$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultLogger-get_beta_summary_df"></a>}}
\if{latex}{\out{\hypertarget{method-ResultLogger-get_beta_summary_df}{}}}
\subsection{Method \code{get_beta_summary_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultLogger$get_beta_summary_df(
  spatial_labels = NULL,
  temporal_labels = NULL,
  feature_labels = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultLogger-get_iteration_betas_tensor"></a>}}
\if{latex}{\out{\hypertarget{method-ResultLogger-get_iteration_betas_tensor}{}}}
\subsection{Method \code{get_iteration_betas_tensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultLogger$get_iteration_betas_tensor(
  spatial_labels,
  temporal_labels,
  feature_labels
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultLogger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResultLogger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultLogger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
