% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample_norm_multivariate}
\alias{sample_norm_multivariate}
\title{Sample a tensor of random values from a normal multivariate distribution}
\usage{
sample_norm_multivariate(mean_vec, precision_lower_tri)
}
\description{
The sampling use a tensor of
}
\keyword{internal}
