% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_spatiotemporal_data}
\alias{simulate_spatiotemporal_data}
\title{Simulate Spatiotemporal Data Using Kernel Covariances.}
\usage{
simulate_spatiotemporal_data(
  nb_locations,
  nb_time_points,
  nb_spatial_dimensions,
  spatial_scale,
  time_scale,
  spatial_covariates_means,
  temporal_covariates_means,
  spatial_kernel,
  temporal_kernel,
  noise_variance_scale
)
}
\arguments{
\item{nb_locations}{Integer: Number of spatial locations}

\item{nb_time_points}{Integer: Number of time points}

\item{nb_spatial_dimensions}{Integer: Number of spatial dimensions}

\item{spatial_scale}{Numeric: Spatial scale}

\item{time_scale}{Numeric: Time scale}

\item{spatial_covariates_means}{Vector: Spatial covariates means}

\item{temporal_covariates_means}{Vector: Temporal covariates means}

\item{spatial_kernel}{Kernel: Spatial kernel}

\item{temporal_kernel}{Kernel: Temporal kernel}

\item{noise_variance_scale}{Numeric: Noise variance scale}
}
\value{
A list containing 4 dataframes:
     - `data_df` contains the response variable and the covariates
     - `spatial_positions_df` contains the spatial locations and their coordinates
     - `temporal_positions_df` contains the time points and their coordinates
     - `beta_df` contains the true beta coefficients
}
\description{
Simulate Spatiotemporal Data Using Kernel Covariances.
}
