% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_ops.R
\name{TensorOperator}
\alias{TensorOperator}
\title{R6 singleton that contains the configuration for the tensor backend}
\description{
Tensor backend configuration and methods for all the tensor operations
in BKTR
}
\section{Super class}{
\code{\link[R6P:Singleton]{R6P::Singleton}} -> \code{TensorOperator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TensorOperator-new}{\code{TensorOperator$new()}}
\item \href{#method-TensorOperator-set_params}{\code{TensorOperator$set_params()}}
\item \href{#method-TensorOperator-get_default_jitter}{\code{TensorOperator$get_default_jitter()}}
\item \href{#method-TensorOperator-tensor}{\code{TensorOperator$tensor()}}
\item \href{#method-TensorOperator-is_tensor}{\code{TensorOperator$is_tensor()}}
\item \href{#method-TensorOperator-eye}{\code{TensorOperator$eye()}}
\item \href{#method-TensorOperator-ones}{\code{TensorOperator$ones()}}
\item \href{#method-TensorOperator-zeros}{\code{TensorOperator$zeros()}}
\item \href{#method-TensorOperator-rand}{\code{TensorOperator$rand()}}
\item \href{#method-TensorOperator-randn}{\code{TensorOperator$randn()}}
\item \href{#method-TensorOperator-randn_like}{\code{TensorOperator$randn_like()}}
\item \href{#method-TensorOperator-arange}{\code{TensorOperator$arange()}}
\item \href{#method-TensorOperator-rand_choice}{\code{TensorOperator$rand_choice()}}
\item \href{#method-TensorOperator-kronecker_prod}{\code{TensorOperator$kronecker_prod()}}
\item \href{#method-TensorOperator-khatri_rao_prod}{\code{TensorOperator$khatri_rao_prod()}}
\item \href{#method-TensorOperator-clone}{\code{TensorOperator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-new"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$new(fp_type = "float64", fp_device = "cpu")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-set_params}{}}}
\subsection{Method \code{set_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$set_params(fp_type = NULL, fp_device = NULL, seed = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-get_default_jitter"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-get_default_jitter}{}}}
\subsection{Method \code{get_default_jitter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$get_default_jitter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-tensor"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-tensor}{}}}
\subsection{Method \code{tensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$tensor(tensor_data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-is_tensor"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-is_tensor}{}}}
\subsection{Method \code{is_tensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$is_tensor(tensor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-eye"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-eye}{}}}
\subsection{Method \code{eye()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$eye(eye_dim)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-ones"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-ones}{}}}
\subsection{Method \code{ones()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$ones(tsr_dim)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-zeros"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-zeros}{}}}
\subsection{Method \code{zeros()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$zeros(tsr_dim)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-rand"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-rand}{}}}
\subsection{Method \code{rand()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$rand(tsr_dim)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-randn"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-randn}{}}}
\subsection{Method \code{randn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$randn(tsr_dim)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-randn_like"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-randn_like}{}}}
\subsection{Method \code{randn_like()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$randn_like(input_tensor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-arange"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-arange}{}}}
\subsection{Method \code{arange()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$arange(start, end)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-rand_choice"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-rand_choice}{}}}
\subsection{Method \code{rand_choice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$rand_choice(choices_tsr, nb_sample, use_replace = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-kronecker_prod"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-kronecker_prod}{}}}
\subsection{Method \code{kronecker_prod()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$kronecker_prod(a, b)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-khatri_rao_prod"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-khatri_rao_prod}{}}}
\subsection{Method \code{khatri_rao_prod()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$khatri_rao_prod(a, b)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TensorOperator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TensorOperator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
