% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_ops.R
\name{TensorOperator}
\alias{TensorOperator}
\title{R6 singleton that contains the configuration for the tensor backend}
\description{
Tensor backend configuration and methods for all the tensor operations
in BKTR
}
\keyword{internal}
\section{Super class}{
\code{\link[R6P:Singleton]{R6P::Singleton}} -> \code{TensorOperator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TensorOperator$new()}}
\item \href{#method-set_tensor_type}{\code{TensorOperator$set_tensor_type()}}
\item \href{#method-set_device_type}{\code{TensorOperator$set_device_type()}}
\item \href{#method-new_tensor}{\code{TensorOperator$new_tensor()}}
\item \href{#method-zeros}{\code{TensorOperator$zeros()}}
\item \href{#method-ones}{\code{TensorOperator$ones()}}
\item \href{#method-eye}{\code{TensorOperator$eye()}}
\item \href{#method-arange}{\code{TensorOperator$arange()}}
\item \href{#method-kronecker_prod}{\code{TensorOperator$kronecker_prod()}}
\item \href{#method-khatri_rao_prod}{\code{TensorOperator$khatri_rao_prod()}}
\item \href{#method-clone}{\code{TensorOperator$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$new(
  tensor_dtype = torch::torch_float64(),
  tensor_device = "cpu"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_tensor_type"></a>}}
\if{latex}{\out{\hypertarget{method-set_tensor_type}{}}}
\subsection{Method \code{set_tensor_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$set_tensor_type(tensor_dtype)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_device_type"></a>}}
\if{latex}{\out{\hypertarget{method-set_device_type}{}}}
\subsection{Method \code{set_device_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$set_device_type(tensor_device)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new_tensor"></a>}}
\if{latex}{\out{\hypertarget{method-new_tensor}{}}}
\subsection{Method \code{new_tensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$new_tensor(tensor_dim)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-zeros"></a>}}
\if{latex}{\out{\hypertarget{method-zeros}{}}}
\subsection{Method \code{zeros()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$zeros(tensor_dim)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ones"></a>}}
\if{latex}{\out{\hypertarget{method-ones}{}}}
\subsection{Method \code{ones()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$ones(tensor_dim)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eye"></a>}}
\if{latex}{\out{\hypertarget{method-eye}{}}}
\subsection{Method \code{eye()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$eye(eye_dim)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-arange"></a>}}
\if{latex}{\out{\hypertarget{method-arange}{}}}
\subsection{Method \code{arange()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$arange(start, end)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kronecker_prod"></a>}}
\if{latex}{\out{\hypertarget{method-kronecker_prod}{}}}
\subsection{Method \code{kronecker_prod()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$kronecker_prod(a, b)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-khatri_rao_prod"></a>}}
\if{latex}{\out{\hypertarget{method-khatri_rao_prod}{}}}
\subsection{Method \code{khatri_rao_prod()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$khatri_rao_prod(a, b)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TensorOperator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
